% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoData.R
\docType{data}
\name{Bullfrogs}
\alias{Bullfrogs}
\title{Reproductive success of male bullfrogs}
\format{data.frame with 38 observations and 8 variables}
\source{
These data were obtained from Table 1 in Howard (1979): \url{http://www.jstor.org/stable/2460219}.
}
\usage{
Bullfrogs
}
\description{
Raw data from the Howard (1979) study on sexual selection on body size in male \emph{Rana catesbeiana} bullfrogs during 1976.  These data were transcribed from Table 1 in Howard (1979) and then formatted to be used in tremors. \code{Bullfrogs} contains the following variables:

\describe{
\item{\code{BodySize.mm}}{Body size of the individual, based on snout-ischium length in millimeters}
\item{\code{w.TotalCops}}{Relative fitness for mating success (sexual selection), based on the total number of copulations.}
\item{\code{w.TotalZygotes}}{Relative fitness for fertility per mate (natural selection), based on the total number of zygotes produced.}
\item{\code{w.TotalHatchlings}}{Relative fitness for offspring survivorship (natural selection), based on the total number of hatchlings surviving.}
\item{\code{W.TotalCops}}{Absolute fitness for mating success (sexual selection), based on the total number of copulations.}
\item{\code{W.TotalZygotes}}{Absolute fitness for fertility per mate (natural selection), based on the total number of zygotes produced.}
\item{\code{W.TotalHatchlings}}{Absolute fitness for offspring survivorship (natural selection), based on the total number of hatchlings surviving.}
\item{\code{Ind}}{Individual identification number. Note: one individual had a multi-number ID (i.e., 3-7) and another was not assigned a specific ID and was designated as "NA".}
}
}
\examples{
# Load the data
data(Bullfrogs)

# Look at the structure of the data.frame
str(Bullfrogs)

# Run a linear regression with w.TotalCops as the response and the morphological traits as the predictors
lm(wTotalZygotes ~ BodySize.mm, data=Bullfrogs)
}
\references{
Howard RD. 1979. Estimating reproductive success in natural populations. \emph{American Naturalist} 114(2):221-231. \url{http://dx.doi.org/10.1086/667586}
}
\keyword{datasets}

