% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoDatasets.R
\docType{data}
\name{BumpusFemales}
\alias{BumpusFemales}
\title{Bumpus - Females only}
\format{data.frame with 49 observations and 13 variables}
\source{
A spreadsheet of the morphological data was obtained from the Field Museum: \url{http://www.fieldmuseum.org/science/blog/hermon-bumpus-and-house-sparrows}, with measurements in inches converted to millimeters.
}
\usage{
BumpusFemales
}
\description{
Raw data from the \code{Bumpus} (1899) study on morphological selection in sparrows. Humerus length, femur length, tibiotarsus length, skull width, and sternum keel length were originally in inches, but are converted to millimeters in this version in order to make the units of scale congruent amongst the traits. Bumpus (1899) analyzed the females and males separately, so this subset includes only the data from the females. \code{BumpusFemales} contains the following variables:

\describe{
		\item{\code{w}}{relative fitness, based on survival for only females pooled together.}
		\item{\code{W}}{absolute fitness, based on survival for only females pooled together.}
		\item{\code{TotalLength.mm}}{length from the tip of the beak to the tip of the tail in millimeters.}
		\item{\code{AlarExtent.mm}}{length of the wing span, from the tip of the left wing to the tip of the right wing in millimeters.}
		\item{\code{Weight.g}}{total weight in grams.}
		\item{\code{SkullL.mm}}{length of the skull, from the tip of the beak to the occiput in millimeters.}
		\item{\code{HumerusL.mm}}{length of the humerus (upper arm bone) in millimeters.}
		\item{\code{FemurL.mm}}{length of the femur (thigh bone) in millimeters.}
		\item{\code{TibioTarL.mm}}{length of the tibiotarsus (lower leg bone) in millimeters.}
		\item{\code{SkullW.mm}}{width of the skull, from the left postorbital bone to the right postorbital bone, in millimeters.}
		\item{\code{SternumL.mm}}{length of the keel of the sternum in millimeters.}
		\item{\code{Sex}}{sex of the individual (1 = male, 0 = female).}
		\item{\code{Age.Group}}{ontogenetic age group (1 = adult, 0 = young, NA = not available).}
		}
}
\examples{
# Load the data
data(BumpusFemales)

# Look at the structure of the data.frame
str(BumpusFemales)

# Run a linear regression with w as the response and the morphological traits as the predictors
lm(w ~ ., data=BumpusFemales[,c(1,3:11)])
}
\references{
Bumpus, H.C. 1899. The elimination of the unfit as illustrated by the introduced sparrow, \emph{Passer domesticus}. \emph{Biol. Lectures, Woods Hole Marine Biol. Station}: 209-226.
}
\seealso{
\code{\link{Bumpus}, \link{BumpusMales}}
}
\keyword{datasets}

