% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoDatasets.R
\docType{data}
\name{CichlidEjagham}
\alias{CichlidEjagham}
\title{Cameroon crater lake cichlids from the Ejagham lake}
\format{data.frame with 523 observations and 11 variables}
\source{
These data were obtained from the "fitness dataset" tab of the "Complete Dataset" Excel spreadsheet that is available on Dryad: \url{http://datadryad.org/resource/doi:10.5061/dryad.rn30d}, and are associated with Martin (2012).
}
\usage{
CichlidEjagham
}
\description{
Raw data from the Martin (2012) study on morphological selection in Cameroon crater lake cichlids.  These data were downloaded from Dryad and then formatted for use in this R package. \code{CichlidEjagham} is a subset of the Martin (2012) data that includes the fish from the Ejagham lake, and contains the following variables:

\describe{
\item{\code{w}}{relative fitness, based on relative growth rates averaged from multiple measurements of the spacing between scale circuli of fishes from Ejagham.}
\item{\code{W}}{absolute fitness, based on relative growth rates averaged from multiple measurements of the spacing between scale circuli of fishes from Ejagham.}
\item{\code{ascendingProcess_cm}}{distance "from the dorsal tip of the ascending process of the premaxilla to the tip of the most anterior tooth on the dentigerous arm of the premaxilla" (in centimeters).}
\item{\code{bodyDepth_cm}}{the distance "from the insertion of the first dorsal fin ray to the ventral surface, perpendicular to the major axis of the fish" (in centimeters).}
\item{\code{orbitDiameter_cm}}{average of the diameters of the orbit (eye socket), taken from the horizontal and vertical axes of the fish (in centimeters).}
\item{\code{headDepth_cm}}{distance "from the insertion of the epaxial muscle on the dorsal surface of the neurocranium to the ventral surface, perpendicular to the major axis of the fish" (in centimeters).}
\item{\code{jawLength_cm}}{distance "from the center of the protruding quadrate-articular joint on the external jaw line to the tip of the most anterior tooth on the mandible" (in centimeters).}
\item{\code{rowID}}{integer to identify the row number for each observation.}
\item{\code{fishID}}{individual identification number for each fish.}
\item{\code{lake}}{lake that the fish were collected from.}
\item{\code{speciesDiagnosticColoration}}{assignment of species based on species-diagnostic-melanin-coloration. Individuals that could not be accurately identified to species were labeled as "ambiguous".}
}
}
\examples{
# Load the data
data(CichlidEjagham)

# Look at the structure of the data.frame
str(CichlidEjagham)

# Run a linear regression with w as the response and the morphological traits as the predictors
lm(w ~ ., data=CichlidEjagham[,1:5])
}
\references{
Martin CH. 2012. Weak disruptive selection and incomplete phenotypic divergence in two classic examples of sympatric speciation: Cameroon crater lake cichlids. \emph{American Naturalist} 180:E90-109. \url{http://dx.doi.org/10.1086/667586}
}
\seealso{
\code{\link{CichlidBarombi}}
}
\keyword{datasets}

