% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_outcome.R
\name{combine_outcome}
\alias{combine_outcome}
\title{Combine GWAS outcome of each batch}
\usage{
combine_outcome(model, chr,
  path = "/Users/boxi lin/Desktop/sickkids/psa/result/gwas_cov/")
}
\arguments{
\item{model}{e.g. "baseline" or "log"}

\item{chr}{indicator for which chromosome}

\item{path}{path until the model name}
}
\description{
Combine GWAS outcome of each batch
}
