% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_analysis.R
\name{compare_analysis}
\alias{compare_analysis}
\title{Comparison of PSA and Benchmark Methods}
\usage{
compare_analysis(X)
}
\arguments{
\item{X}{a data matrix}
}
\value{
a list of data matrix \code{X} and outcomes of PSA-S, PSA-O, PCA, log-ratio PCA,
and power transform PCA with power 1/2.
}
\description{
A wrapper function of PSA and benchmark methods for convenience
}
\seealso{
\code{\link[=psa]{psa()}}, \code{\link[=comp_pca]{comp_pca()}}, \code{\link[=comp_apca]{comp_apca()}}, \code{\link[=power_pca]{power_pca()}}
}
