% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionDifferentials.R
\name{dCov}
\alias{dCov}
\title{Estimating linear and nonlinear selection differentials}
\usage{
dCov(w, z)
}
\arguments{
\item{\code{w}}{Relative fitness.}

\item{\code{z}}{Phenotypic trait(s). Character values are not accepted.}
}
\value{
\code{dCov} returns a vector of numeric values.
}
\description{
\code{dCov} estimates the linear and nonlinear selection differentials for one or more phenotypic traits, based on descriptions in Phillips and Arnold (1989). Estimations are based on phenotypic traits that have been normalized to a mean of zero and unit variance. If one phenotypic trait is input, linear and quadratic selection differentials will be output. If more than one phenotypic trait is input, correlational selection differentials will also be output.
}
\section{Output}{
  \code{dCov} is the selection differential calculated from the covariance between the relative fitness, \code{w}, and each phenotypic trait \code{z}, using the approach described in Table 1 of Phillips and Arnold (1989).
}
\examples{
data(BumpusMales)

dCov(BumpusMales$w, BumpusMales[,3:11])
}
\references{
Lande R, Arnold SJ. 1983. The measurement of selection on correlated characters. \emph{Evolution} 37(6): 1210-1226. \url{http://www.jstor.org/stable/2408842}
}

