% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoDatasets.R
\docType{data}
\name{DesertPlants}
\alias{DesertPlants}
\title{Phenotypic selection in winter annual plants from the Sonoran Desert}
\format{data.frame with 913 observations and 9 variables}
\source{
These data were obtained from the "KimballetalData.xlsx" data file available on Dryad: \url{http://dx.doi.org/10.5061/dryad.c8c58}, and are associated with Kimball et al. (2013)
}
\usage{
DesertPlants
}
\description{
Standardized data from the Kimball et al. (2013) study on phenotypic selection in four winter annual plant species from the Sonoran Desert that identified a trade-off between relative growth rate and water use efficiency. Traits were standardized by "substracting the mean and dividing by the standard deviation". These data were downloaded from Dryad and then formatted for use in this R package. \code{DesertPlants} consists of the following variables:

\describe{
		\item{\code{w}}{relative fitness, based on total plant biomass (\code{RelFitness}), which was used as a proxy for seed production.}
		\item{\code{stdRGR}}{standardized relative growth rate.}
		\item{\code{stdSLA}}{standardized specific leaf area, which was calculated as leaf area/dry mass of the leaf.}
		\item{\code{stdRMR}}{standardized root mass ratio, which was calculated as root dry mass/total dry mass.}
		\item{\code{stdN}}{standardized leaf Nitrogen content.}
		\item{\code{stdDELTA}}{standardized integrated water-use efficiency over the lifetime of the leaf, based on carbon isotope ratios.}
		\item{\code{Site}}{collection site within the Sonoran Desert. TH = University of Arizona's Desert Laboratory at Tumamoc Hill; cooler and wetter site. OPNM = Organ Pipe National Monument; warmer and drier site.}
		\item{\code{Species}}{species of plant. STMI = \emph{Stylocline micropoides}, ERLA = \emph{Eriophyllum lanosum}, PERE = \emph{Pectocarya recurvata}, and ERTE = \emph{Erodium texanum}.}
		\item{\code{ID}}{individual identification number within each treatment/species combination.}
}
}
\examples{
# Load the data
data(DesertPlants)

# Look at the structure of the data.frame
str(DesertPlants)

# Run a linear regression with w as the response and the morphological traits as the predictors
lm(w ~ ., data = DesertPlants[,1:6])
}
\references{
Kimball S, Gremer JR, Huxman TE, Venable DL, Angert AL (2013) Phenotypic selection favors missing trait combinations in coexisting annual plants. \emph{The American Naturalist} 182(2): 191-207.  \url{http://dx.doi.org/10.1086/671058}
}
\keyword{datasets}

