% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionDifferentials.R
\name{differentials}
\alias{differentials}
\title{Estimate phenotypic selection differentials}
\usage{
differentials(w, z, method = c(1,2,3,4, "all"), standardize = TRUE, ...)
}
\arguments{
\item{\code{w}}{Relative fitness.}

\item{\code{z}}{Phenotypic trait(s). Character values are not accepted.}

\item{\code{method}:}{method to estimate the selection differential. 1 = covariance of relative fitness to the trait; 2 = differences in mean, variance, and covariance before and after selection; 3 = matrix algebra approach of phenotypic distributions before and after selection; 4 = ordinary least-squares regression of relative fitness against the trait; "all" = use all of the methods to produce multiple estimates.}

\item{\code{standardize}}{Indicate whether phenotypic trait data should be standardized to a mean of zero and unit variance.}
}
\value{
\code{differentials} returns a data.frame
}
\description{
\code{differentials} allows the user to evaluate how different methods for calculating phenotypic section differentials influence the output and, thus, the interpretation of indirect selection on phenotypic traits.
}
\section{Output}{
  \code{dCov} is the selection differential calculated from the covariance between the relative fitness, \code{w}, and each phenotypic trait \code{z} (Lande and Arnold 1983).

  \code{dBeforeAfter} is the selection differential calculated as  as described in Brodie et al. (1995).

 \code{dMatrix} is the selection differential calculated using matrix algebra (e.g., equation 13b in Lande and Arnold (1983))

  \code{dReg} is the selection differential calculated as the partial regression coefficient of relative fitness, \code{w}, against each individual phenotypic trait through univariate linear regressions.

  \code{dCovScale} calculates the selection differential using the equation in \code{dCovScale}, but uses z data that are standardized to a mean of zero and unit variance.
}
\examples{
data(BumpusMales)

differentials(BumpusMales$w, BumpusMales[,3:11], "all")
}
\references{
Brodie III ED, Moore AJ, Janzen FJ. 1995. Visualizing and quantifying natural selection. \emph{Trends in Ecology and Evolution} 10(8): 313-318. \url{http://www.sciencedirect.com/science/article/pii/S016953470089117X}

Lande R, Arnold SJ. 1983. The measurement of selection on correlated characters. \emph{Evolution} 37(6): 1210-1226. \url{http://www.jstor.org/stable/2408842}
}

