% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoData.R
\docType{data}
\name{Flowers}
\alias{Flowers}
\title{Flower morphology Morrenia brachtephana}
\format{data.frame with 125 observations and 12 variables}
\source{
These data were obtained from "Phenotypic Selection Analysis" data file available on Dryad: \url{http://dx.doi.org/10.5061/dryad.kq72j}, and are associated with Baranzelli et al. (2014).
}
\usage{
Flowers
}
\description{
Raw data from the Baranzelli et al. (2014) study on morphological selection in the flowers of \emph{Morrenia brachtephana}.  These data were downloaded from Dryad and then formatted to be used in tremors. \code{Flowers} consists of the following variables:

\describe{
		\item{\code{wm}}{relative male fitness, based on the average number of pollinaria exported per flower from all the flowers sampled for each individual (\code{Pollinarium.export}).}
		\item{\code{wf}}{relative female fitness, based on the average number of pollinia received per flower from the total number collected per individual (\code{pollinia.import}).}
		\item{\code{Wm}}{absolute male fitness, based on the average number of pollinaria exported per flower from all the flowers sampled for each individual (\code{Pollinarium.export}).}
		\item{\code{Wf}}{absolute female fitness, based on the average number of pollinia received per flower from the total number collected per individual (\code{pollinia.import}).}
		\item{\code{At3}}{width of the corona, a trait related to the visual attraction of the flower.}
		\item{\code{Gy2}}{width of the inner guide rail, a trait related to pollinaria transfer.}
		\item{\code{Gy3}}{width of the inner guide rail at the level of the upper edges, a trait related to pollinaria transfer.}
		\item{\code{Gy4}}{length of the inner guide rail, a trait related to pollinaria transfer.}
		\item{\code{Gy5}}{length of the outer guide rail, a trait related to pollinaria transfer.}
		\item{\code{Co3}}{length of the corpusculum, a measure of the pollinarium.}
		\item{\code{Ca3}}{length of the caudicles, a measure of the pollinarium.}
		\item{\code{individual}}{individual identification number.}
}
}
\examples{
# Load the data
data(Flowers)

# Look at the structure of the data.frame
str(Flowers)

# Run a linear regression with wm as the response and the morphological traits as the predictors
lm(wm ~ ., data = Flowers[,5:11])
}
\references{
Baranzelli MC, Sersic AN, Cocucci AA. 2014. The search for Pleiades in trait constellations: functional integration and phenotypic selection in the complex flowers of \emph{Morrenia brachystephana} (Apocynaceae). \emph{Journal of Evolutionary Biology} 27(4): 724-736. \url{http://dx.doi.org/10.1111/jeb.12341}
}
\keyword{datasets}

