% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoDatasets.R
\docType{data}
\name{Galls}
\alias{Galls}
\title{Gall size}
\format{data.frame with 14283 observations and 6 variables}
\source{
These data were obtained from the Excel spreadsheet that is available on Dryad: \url{http://datadryad.org/resource/doi:10.5061/dryad.1js1n}, and are associated with Egan et al. (2011).
}
\usage{
Galls
}
\description{
Raw data from the Egan et al. (2011) study on phenotypic selection on gall size produced by \emph{Belonocnema treatae} on a host tree (\emph{Quercus fusiformis}).  These data were downloaded from Dryad and then formatted for use in this R package. \code{Galls} consists of the following variables:

\describe{
\item{\code{w}}{relative fitness, based on the emergence/survivorship of the host-specific gall-former \emph{Belonocnema treatae} (\code{BtreataeEmergence}) on it host plant (the plateau live oak \emph{Quercus fusiformis}).}
\item{\code{W}}{absolute fitness, based on the emergence/survivorship of the host-specific gall-former \emph{Belonocnema treatae} (\code{BtreataeEmergence}) on it host plant (the plateau live oak \emph{Quercus fusiformis}).  1 = emerged/survived, 0 = no emergence.}
\item{\code{gallSizeDiameter_mm}}{diameter of the galls produced by \emph{B. treatae} on its host plant, in units of millimeters and measured to the nearest 0.01 mm.}
\item{\code{year}}{year that the data were collected (either 2004 or 2008)}
\item{\code{tree}}{individual identification of the tree that was sampled. Each tree was treated as a separate population.}
\item{\code{bagPerReplicate}}{identication number of the bag/enclosure that was treated as a blocking factor within each tree.}
}
}
\examples{
# Load the data
data(Galls)

# Look at the structure of the data.frame
str(Galls)

# Run a linear regression with wr as the response and the phenotypic traits as the predictors
lm(wr ~ ., data = Galls[,3])
}
\references{
Egan SP, Hood GR, Ott JR. 2011. Natural selection on gall size: variable contributions of individual host plants to population-wide patterns. \emph{Evolution} 65(12): 3543-3557. \url{doi:10.1111/j.1558-5646.2011.01396.x}
}
\keyword{datasets}

