% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strata_breaks.R
\name{get_strata_breaks}
\alias{get_strata_breaks}
\title{Stratify based upon the propensity scores.}
\usage{
get_strata_breaks(ps, n_strata = 5, labels = LETTERS[1:n_strata])
}
\arguments{
\item{ps}{vector containing the propensity scores.}

\item{n_strata}{number of strata.}

\item{labels}{labels for the strata.}
}
\value{
a list with two elements:
\describe{
\item{breaks}{a numeric vector with the breaks returned from [stats::quantile()].}
\item{labels}{a data frame with four columns: strata, xmin (lower bound for the stratum), 
          xmax (upper bound for the stratum), and xmid (midpoint of the stratum).} 
}
}
\description{
Stratify based upon the propensity scores.
}
