% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmean.R
\name{ggbmean}
\alias{ggbmean}
\title{Plot the backwards mean}
\usage{
ggbmean(result, plot = TRUE)
}
\arguments{
\item{result}{Output of psoa() pssa() or similar.}

\item{plot}{If TRUE, the makes a parallel coordinates plot using \code{ggplot2}. Otherwise returns the backward mean.}
}
\description{
I \emph{think} this is equivalent to the zero-dimensional result, i.e. \verb{result$vertices["r=1"]]}
}
