% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_modes.R
\name{ggscores_pairs}
\alias{ggscores_pairs}
\title{Plot scores from two merges}
\usage{
ggscores_pairs(result, ..., dimensions, mapping = NULL)
}
\arguments{
\item{result}{Output from psoa or pssa}

\item{...}{Passed to tidy_scores}

\item{dimensions}{Passed to tidy_scores. Note dimensions is a misnomer - it is actually the number of vertices.}

\item{mapping}{Create using \code{ggplot2::aes()}, do not specify \code{x} and \code{y} aesthetics - these are created from \code{dimensions} (possibly hacky, expoiting the fact that aes() creates a list).}
}
\description{
Scatter plot of scores. Along the border of the plot the component names making up the vertices involved are printed.
The top border has the components that make up the positive-direction vertex of the merge specified by \code{y} in \code{mapping}.
The bottom border has the components that make up the negative-direction vertex of the merge.

For PSSA scores depend on the ratio of mass in each of the vertices in the merge, and also on the total mass in both vertices (this could potentially be removed later for ).
}
