% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionGradients.R
\name{glam}
\alias{glam}
\title{Calculating gradients based on the Lande-Arnold Method}
\usage{
glam(w, z, wType=c("gaussian", "binomial"), JS = FALSE, prep = TRUE, st= NULL, RE = NULL)
}
\arguments{
\item{\code{fitness}}{Fitness measure. Gaussian fitness types should be use relative which, which is calculated as the absolute fitness for each individual \code{W(z)} divided by the mean absolute fitness \code{W}. Binomial fitness types should use the absolute fitness measures (e.g., 0 = failed, 1 = survived)}

\item{\code{z}}{Phenotypic traits.}

\item{\code{fitType}}{Type of distribution for relative fitness \code{w}. Option to either "gaussian" or "binomial".}

\item{\code{JS}}{Janzen and Stern (1998) correction factor for make logistic regression coefficients congruent with linear regression coefficients for estimating multivariate selection. Default is \code{FALSE}. Setting \code{JS = TRUE} applies the Janzen and Stern correction factor to the logistic regression coefficients.}

\item{\code{prep}}{Option to scale the phenotypic trait data (\code{z}) to a mean of zero and unit variance in preparation for running regression models. Default is set to \code{TRUE}, with the assumption that \code{z} are the raw data for the morphological traits.}

\item{\code{st}}{Option to standardize the regression coefficients by either the mean or the standard deviation of \code{z}.}

\item{\code{RE}}{Random effects of the model, if applicable.}
}
\value{
The function returns an object of classes "\code{glam}", "\code{lm}", and "\code{glm}."
}
\description{
\code{glam} is used to fit generalized linear models, specified by error distributions and link functions as denoted by \code{family}, using approaches based on the quantitative framework established by Lande and Arnold (1983). Statistical methods are based on \code{glm} for linear models and \code{glmer} for linear mixed effects models. An option to employ the Janzen and Stern (1998) correction factor for logistic regression models is available via \code{JS = TRUE}. Model formulae are constructed such that regression coefficients and standard errors for quadratic terms do NOT need to be doubled (Stinchcombe et al. 2008)
}
\details{
The Lande-Arnold Method is based on the 1983 paper by Russell Lande and Stevan Arnold, entitled "The measurement of selection on correlated characters". Their method involves applying ordinary least-squares (OLS) regression to estimate selection gradients.
}
\section{Value}{
 \code{GL}
}

\section{Warning}{
 \strong{These analyses are currently only available for longitudinal data}. Selection gradients for cross-sectional data must be evalated using matrix algebra rather than OLS regressions (Lande and Arnold 1983).
}
\examples{
# use the BumpusMales data set from the tremors package
data(BumpusMales)

# Define the input
w <- BumpusMales$w
z <- BumpusMales[,3:11]

# Calculate the selection gradients using glam
mod1 <- glam(w, z, "gaussian")

# Review the summary statistics
summary.glam(mod1)
}
\references{
Janzen FJ, Stern HL. 1998. Logistic regression for empirical studies of multivariate selection. \emph{Evolution} 52(6): 1564-1571. \url{http://www.jstor.org/stable/2411330?seq=1#page_scan_tab_contents}

Lande R, Arnold SJ. 1983. The measurement of selection on correlated characters. \emph{Evolution} 37(6): 1210-1226. \url{http://www.jstor.org/stable/2408842}

Stinchcombe JR, Agrawal AF, Hohenlohe PA, Arnold SJ, Blows MW. 2008. Estimating nonlinear selection gradients using quadratic regression coefficients: double or nothing? \emph{Evolution} 62(9): 2435-2440. \url{http://onlinelibrary.wiley.com/doi/10.1111/j.1558-5646.2008.00449.x/abstract}
}
\seealso{
\code{glamx}, \code{glm}, \code{lm}, \code{summary.blam}, \code{glmer}
}

