% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionGradients.R
\name{glamx}
\alias{glamx}
\title{Comparison of selection gradients with different trait choices}
\usage{
glamx(fitness, z, method=c("linear", "nonlinear", "both"))
}
\arguments{
\item{\code{fitness}}{Fitness measure. Gaussian fitness types should be use relative fitness, which is calculated as the absolute fitness for each individual \code{W(z)} divided by the mean absolute fitness \code{W}. Binomial fitness types should use the absolute fitness measures (e.g., 0 = failed, 1 = survived). *Note* Only gaussian fitness measures are currently accepted in this version of the function.}

\item{\code{z}}{Phenotypic traits.}

\item{\code{method}}{Choice of whether analyses will be output for linear selection, nonlinear selection, or both.}
}
\value{
The function returns an object of classes "\code{glam}", "\code{lm}", and "\code{glm}."
}
\description{
\code{glamx} is used to fit generalized linear models, specified by error distributions and link functions as denoted by \code{family}, using approaches based on the quantitative framework established by Lande and Arnold (1983). Statistical methods are based on \code{glm} for linear models and \code{glmer} for linear mixed effects models. An option to employ the Janzen and Stern (1998) correction factor for logistic regression models is available via \code{JS = TRUE}. Model formulae are constructed such that regression coefficients and standard errors for quadratic terms do NOT need to be doubled (Stinchcombe et al. 2008). *Note* Only the OLS is currently supported in this version of the function.
}
\details{
The Lande-Arnold Method is based on the 1983 paper by Russell Lande and Stevan Arnold, entitled "The measurement of selection on correlated characters". Their method involves applying ordinary least-squares (OLS) regression to estimate selection gradients.
}
\section{Value}{
 \code{GL}
}

\section{Warning}{
 \strong{These analyses are currently only available for longitudinal data}. Selection gradients for cross-sectional data must be evalated using matrix algebra rather than OLS regressions (Lande and Arnold 1983).
}
\examples{
# use the BumpusMales data set 
data(BumpusMales)

# Define the input
fitness <- BumpusMales$w
z <- BumpusMales[,3:11]

# Calculate the selection gradients using glam
mod1 <- glam(fitness, z, method = "linear")

}
\references{
Lande R, Arnold SJ. 1983. The measurement of selection on correlated characters. \emph{Evolution} 37(6): 1210-1226. \url{http://www.jstor.org/stable/2408842}
}
\seealso{
\code{glam}, \code{lm},
}

