% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionGradients.R
\name{gradients}
\alias{gradients}
\title{Estimating phenotypic selection gradients}
\usage{
gradients(w, z, method = c(1,2, "all"), centered = TRUE, scaled = TRUE, printmod = FALSE, ...)
}
\arguments{
\item{\code{w}}{Relative fitness.}

\item{\code{z}}{Phenotypic trait(s). Character values are not accepted. Input should be raw/unstandardized/un-normalized data, so the data can be normalized correctly for the P star matrix later.}

\item{\code{method}:}{method to estimate the selection differential. 1 = matrix algebra approach of phenotypic distributions before and after selection; 2 = ordinary least-squares regression of relative fitness against the trait; "all" = use all of the methods to produce multiple estimates.}

\item{\code{centered}}{Indicate whether phenotypic trait data should be centered to a mean of zero}

\item{\code{scaled}}{Indicate whether phenotypic trait data should be scaled to unit variance.}
}
\description{
\code{gradients} estimates the linear and nonlinear selection gradients, based on the Lande and Arnold (1983) paper.
}
\details{
Method 2 ordinary least-squares (OLS) regressions are used to estimate selection gradients, based on methods outlined by Lande and Arnold (1983). Quadratic terms have already been coded, so that their regression estimates and standard errors do NOT need to be doubled (Stinchcombe et al. 2008). The matrix algebra approach estimates the linear and nonlinear selection coefficients using equations outlined in Lande and Arnold (1983): linear selection differential (s) is based on eqation 4, linear selection gradient (beta) is equation 17, nonlinear selection differential (C) is equation 13b, and nonlinear selection gradient (gamma) is equation 14a.
}
\examples{

data(BumpusMales)
gradients(BumpusMales$w, BumpusMales[,3:11], "all")
}
\references{
Lande R, Arnold SJ. 1983. The measurement of selection on correlated characters. \emph{Evolution} 37(6): 1210-1226. \url{http://www.jstor.org/stable/2408842}

Stinchcombe JR, Agrawal AF, Hohenlohe PA, Arnold SJ, Blows MW. 2008. Estimating nonlinear selection gradients using quadratic regression coefficients: double or nothing? \emph{Evolution} 62(9): 2435-2440. \url{http://onlinelibrary.wiley.com/doi/10.1111/j.1558-5646.2008.00449.x/abstract}
}
\seealso{
\code{glam}, \code{differentials}
}

