% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.R
\name{gridsearch}
\alias{gridsearch}
\title{Find the best merge by grid search}
\usage{
gridsearch(X, evaluator, testweights = seq(0, 1, length.out = 100), ...)
}
\arguments{
\item{X}{matrix of data points for a simplex with vertices at the basis coordinates.}

\item{evaluator}{A function that evaluates each proposed merge and return a single non-negative value for the quality of the merge. Arguments of \code{X}, \code{v1}, \code{v2}, and \code{w} at least}

\item{testweights}{A vector of weights to try.}

\item{...}{Passed to \code{evaluator}}
}
\value{
The best pair and weight as a single row data frame (data frame for easier row binding and because the v1 and v2 are best stored as integers).
}
\description{
Find the best merge by grid search
}
