% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntensityOfSelection.R
\name{i.AW}
\alias{i.AW}
\title{Intensity of selecion using the selection differential and phenotypic standard deviation AKA standardized selection differentials}
\usage{
i.AW
}
\description{
The total intensity of selection represents the maximum amount of response to selection possible for that sample (Crow 1958). According to Arnold and Wade (1984), selection intensities can be calculated as "selection differentials divided by phenotypic standard deviations before selection", which is equivalent to the standardized selection differential. This metric can be applied to quantify how much directional selection shifts the mean of the phenotypic trait, in units of standard deviations, which is "useful for comparing selective pressures in different populations" (Arnold and Wade 1984). However, there is an assumption of normality if one calculates selection intensities in this fashion. According to Falconer (1989),"The intensity of selection, i, depends only on the proportion of the population included in the selected group and, provided the distribution of phenotypic values is normal, it can be determined from tables of the properties of the normal distribution. If p is the proportion selected, i.e., the proportion of the population falling beyond the point of trunction, and z is the height of the ordinate at the point of truncation, then it follows from mathematical properties of the normal distribution that": (selection differential/phenotypic standard deviation) = standardized selection differential = intensity of selection = z/p, where p is the proportion selected from the sample and z is the phenotypic trait.
}
\references{
Arnold SJ, Wade MJ. 1984. On the measurement of natural and sexual selection: applications. \emph{Evolution} 38(4): 720-734.

Crow JF. 1958. Some possibilities for measuring selection intensities in man. \emph{Human Biology} 30: 1-13. \url{http://www.jstor.org/stable/41449168?seq=1#page_scan_tab_contents}

Falconer DS. 1989. Introduction to Quantitative Genetics. Third Edition. Longman Scientific and Technical. Copublished in the United States with John Wiley and Sons, Inc., New York. p. 192.
}

