% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpportunityForSelection.R
\name{I_total}
\alias{I_total}
\title{Total opportunity for selection}
\usage{
I_total(fitness, type = c("W", "w"))
}
\arguments{
\item{\code{fitness}}{Vector of numeric or integer values that represent the fitness metric or proxy.}

\item{\code{type}}{Type of fitness metric or proxy, either absolute fitness (\code{"W"}) or relative fitness (\code{"w"}).}
}
\value{
\code{I_total} returns a single numeric value.
}
\description{
Calculation of the total opportunity for selection based on the variance in relative fitness: \code{var(W/mean(W))} (Crow 1958, Arnold and Wade 1984, Brodie et al. 1995), where \code{W} = absolute fitness.
}
\references{
Arnold SJ, Wade MJ. 1984. On the measurement of natural and sexual selection applications. \emph{Evolution} 38(4): 720-734. \url{http://www.jstor.org/stable/2408384?seq=1#page_scan_tab_contents}

Brodie ED III, Moore AJ, Janzen FJ. 1995. Visualizing and quantifying natural selection. \emph{TREE} 10(8): 313-318. \url{http://www.sciencedirect.com/science/article/pii/S016953470089117X}

Crow JF. 1958. Some possibilities for measuring selection intensities in man. \emph{Human Biology} 30: 1-13. \url{http://www.jstor.org/stable/41449168?seq=1#page_scan_tab_contents}
}

