% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpportunityForSelection.R
\name{I_traits}
\alias{I_traits}
\title{Partitioning the opportunity for selection due to individual phenotypic traits}
\usage{
I_traits(fitness, z, fitType = c("rel", "abs"), zType = c("long", "cross"))
}
\arguments{
\item{\code{z}}{Phenotypic traits.}

\item{\code{fitness}}{Vector of numeric or integer values that represent the fitness metric or proxy.}

\item{\code{fitType}}{Type of fitness metric or proxy, either absolute fitness (\code{"W"}) or relative fitness (\code{"w"}).}

\item{\code{zType}}{Type of data: "long" = longitudinal, and "cross" = cross-sectional.}
}
\value{
\code{I_traits} returns a matrix of numeric values. 

\itemize{
   \item \code{B:} A vector of selection gradients (normalized to mean of zero and unit variance)
   \item \code{b:} selection differential without imputed variance or weights
   \item \code{q:} Weights to be applied to the phenotypic traits (defaults to 1).
   \item \code{var:} Imputed variance due to missing data, divided by the weight.
   \item \code{s:} A vector of selection differentials (normalized to mean of zero and unit variance)
   \item \code{i:} Additive opportunity for selection for an individual phenotypic trait. 
   \item \code{I.model:} Opportunity of selection, based on the model of phenotypic traits given. 
   \item \code{I.total:} The total opportunity for selection, based on weighted variance of fitness.
}
}
\description{
Calculation of the opportunity for selection based on the method provided by Moorad and Wade (2013), in which the variation in fitness due to a given trait can be calculated as the product of its selection gradient and selection differential. The I_traits function was written using the code provided in the supplementary material of Moorad and Wade (2013), with slight modifications including a slight correction to  s = b * q * unimputed.var rather than s = B * q * unimputed.var. In addition, the original method by Moorad and Wade (2013) estimated the selection differentials using simple linear regressions that could incorporate weights, but \code{I_traits} estimates the selection differentials using the covariance method (Phillips and Arnold 1983) that does not incorporate weights. The \code{I_traits} function also automatically mean-centers the phenotypic traits.
}
\examples{
# load the dataset
data(BumpusMales)
# Calculate the total opportunity for selection
I_total(BumpusMales$W, type = "W")
}
\references{
Lande R, Arnold SJ. 1983. The measurement of selection on correlated characters. \emph{Evolution} 37(6): 1210-1226. \url{http://www.jstor.org/stable/2408842}

Moorad JA, Wade MJ. 2013. Selection gradients, the opportunity for selection, and the coefficient of determination. \emph{The American Naturalist} 181(3): 291-300. \url{http://www.journals.uchicago.edu/doi/abs/10.1086/669158}

Phillips PC, Arnold SJ. 1989. Visualizing multivariate selection. \emph{Evolution} 43(6): 1209-1222. \url{http://www.jstor.org/stable/2409357}
}

