% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoData.R
\docType{data}
\name{LarvalSquirtsHD}
\alias{LarvalSquirtsHD}
\alias{SquirtsHD}
\title{Larvae of sea squirts (\emph{Styela plicata}) in high density}
\format{data.frame with 83 observations and 10 variables}
\source{
These data were obtained from the Excel spreadsheet that is available on Dryad: \url{http://datadryad.org/resource/doi:10.5061/dryad.574r7}, and are associated with Crean et al. (2011).
}
\usage{
LarvalSquirtsHD
}
\description{
Raw data from the Crean et al. (2011) study on phenotypic selection across the metamorphic boundary in \emph{Styela plicata} larvae.  These data were downloaded from Dryad and then formatted to be used in tremors. \code{LarvalSquirtsHD} includes the data from the high density treatment, and consists of the following variables:

\describe{
\item{\code{wr}}{relative fitness based on postmetamorphic survival to reproduction after 4 months in the field (\code{survivalToReprod}) for larvae in the high density treatment.}
\item{\code{ws}}{relative fitness based on postmetamorphic growth, using wet weight (in grams) after 4 months in the field as a proxy (\code{wetWeight4months.g}) for larvae in the high density treatment.}
\item{\code{Wr}}{absolute fitness based on postmetamorphic survival to reproduction after 4 months in the field (\code{survivalToReprod}) for larvae in the high density treatment. 1 = survived, 0 = died.}
\item{\code{Ws}}{absolute fitness based on postmetamorphic growth, using wet weight (in grams) after 4 months in the field as a proxy (\code{wetWeight4months.g}) for larvae in the high density treatment.}
\item{\code{larvalArea.um2}}{area of the larva, "estimated from the average of at least three measurements", in units of squared micrometers.}
\item{\code{hatchTime.mins}}{time to hatching, a premetamorphic trait used as a proxy for development time, in units of minutes.}
\item{\code{settleTime.mins}}{time to settlement, a premetamorphic trait measured every 4 hours, in units of minutes. "Larvae that had not settled within 25 hours posthatch were excluded from their analyses (n = 3)".}
\item{\code{trial}}{one of five replicate runs ("trials") that were run betwen May and October 2009.}
\item{\code{individual}}{individual identification code.}
\item{\code{treatment}}{experimental treatment group (LD = low density, HD = high density).}
}
}
\examples{
# Load the data
data(LarvalSquirtsHD)

# Look at the structure of the data.frame
str(LarvalSquirtsHD)

# Run a linear regression with wr as the response and the phenotypic traits as the predictors
lm(wr ~ ., data = LarvalSquirtsHD[,5:7])
}
\references{
Crean AJ, Monro K, Marshall DJ. 2011. Fitness consequences of larval traits persist across the metamorphic boundary. \emph{Evolution} 65(11): 3079-3089. \url{doi:10.1111/j.1558-5646.2011.01372.x}
}
\seealso{
\code{\link{LarvalSquirts}, \link{LarvalSquirtsLD}}
}
\keyword{datasets}

