% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_bar.R
\name{loading_bar}
\alias{loading_bar}
\title{Loading Bar Plots for PSA}
\usage{
loading_bar(v, max.k = 12)
}
\arguments{
\item{v}{a loading vector}

\item{max.k}{maximum number of elements to display}
}
\value{
a ggplot object
}
\description{
Creating a bar plot summarizing a loading vector
}
