% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchBalance.R
\name{MatchBalance}
\alias{MatchBalance}
\title{Calculates balance statistics for matching.}
\usage{
MatchBalance(df, formu, formu.Y, exact.covs, index.treated, index.control,
  tolerance = 0.25, n.levels = 2, ...)
}
\arguments{
\item{df}{data frame with data to match.}

\item{formu}{the formula used to estimate propensity scores using logistic 
regression.}

\item{formu.Y}{(optional) the formula used to estimate a multiple regression model with
covariates. Used to estimate covariate importance.}

\item{exact.covs}{character vector of covariates to match exactly on. For 
quantitative variables with more than two unique values, matches will 
be made within quintiles.}

\item{index.treated}{a vector of integers corresponding to the rows in \code{df}
for the matched treated units. That is, \code{index.treated[i]} is matched
to \code{index.control[i]}.}

\item{index.control}{a vector of integers corresponding to the rows in \code{df}
for the matched control units. That is, \code{index.treated[i]} is matched
to \code{index.control[i]}.}

\item{tolerance}{the maximum standard deviation the difference between matches
on a quantitative covariate can be to be considered matched.}

\item{n.levels}{maximum number of unique values in a covariate to attempt to
match exactly. If the number of unique values is greater than \code{n.levels},
then the covariate will be cut using quantiles and exact matches will
be estimated within each quantile.}

\item{...}{other parameters passed to \code{\link{Match}}.}
}
\value{
An object of class \code{MatchBalance} with the following elements:
        \describe{
          \item{ps.out}{Results from the logistic regression model for treatment.}
          \item{Y.out}{Results from the multiple regression model for Y.}
          \item{covariate.matched}{a named numeric vector containing the percent
                of observatins with "perfect" match for each covariate.}
          \item{covariate.balance}{a named numeric vector containing the percent
                of covariates with "perfect" match for each observation.}
          \item{covariate.importance}{a data frame with covariate importance.
                Covariate importance is estimated using the sum of the absolute
                t-statistic from both glm calls for treatment and outcome.}
          \item{index.treated}{vector containing observation row number from original
                data frame for treated observations.}
          \item{index.control}{vector containing observatino row number from original
                data frame for control observations.}
        }
}
\description{
Calculates balance statistics for matching.
}
