% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchBalance_to_data_frame.R
\name{MatchBalance_to_data_frame}
\alias{MatchBalance_to_data_frame}
\title{Converts the results MatchBalance to a data.frame.}
\usage{
MatchBalance_to_data_frame(mb)
}
\arguments{
\item{mb}{result from [Matching::MatchBalance()].}
}
\value{
a `data.frame`.
}
\description{
Converts the results MatchBalance to a data.frame.
}
