% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_plot.R
\name{matching_plot}
\alias{matching_plot}
\title{Propensity score matching plot}
\usage{
matching_plot(
  ps,
  treatment,
  outcome,
  index_treated,
  index_control,
  null_hypothesis = 0,
  method = "loess",
  se = TRUE,
  level = 0.95,
  percent_matches = 1,
  colors = c("#fc8d62", "#66c2a5"),
  xlab = "Propensity Score",
  ylab = "Outcome",
  treat_lab = "Treatment",
  plot_points = TRUE
)
}
\arguments{
\item{ps}{vector of propensity scores.}

\item{treatment}{vector of treatment indicators.}

\item{outcome}{vector of outcome values.}

\item{index_treated}{the positions of treated matched pairs.}

\item{index_control}{the positions of control matched pairs.}

\item{null_hypothesis}{the value of the null hypothesis (typically zero). A
horizontal line will be drawn to compare with regression line.}

\item{method}{the method to use for the regression line. Typically loess, gan, or lm.}

\item{se}{Display confidence interval around smooth? (TRUE by default, see level to control.)}

\item{level}{Level of confidence interval to use (0.95 by default).}

\item{percent_matches}{the percentage of matches to plot. Set to one to connect all matches.}

\item{colors}{vector of colors to use for control and treatment, respectively.}

\item{xlab}{label for the x-axis.}

\item{ylab}{label for the y-axis.}

\item{treat_lab}{label for the legend of group colors.}

\item{plot_points}{whether to plot the individual points.}
}
\value{
a ggplot2 expression.
}
\description{
Propensity score matching plot
}
\examples{
if(require(Matching)) {
data(lalonde, package = 'Matching')
lr_out <- glm(treat ~ age + I(age^2) + educ + I(educ^2) + black + 
              hisp + married + nodegr + re74  + I(re74^2) + re75 + I(re75^2) +
              u74 + u75,
			  data = lalonde, 
			  family = binomial(link = 'logit'))
lalonde$ps <- fitted(lr_out)
match_out <- Matching::Match(Y = lalonde$re78,
							 Tr = lalonde$treat,
							 X = lalonde$ps,
							 caliper = 0.1,
							 replace = FALSE,
							 estimand = 'ATE')
matching_plot(ps = lalonde$ps,
			  treatment = lalonde$treat,
			  outcome = log(lalonde$re78 + 1),
			  index_treated = match_out$index.treated,
			  index_control = match_out$index.control)
}
}
