% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionDifferentials.R
\name{multiprod}
\alias{multiprod}
\title{Multiple pair-wise products of data}
\usage{
multiprod(x)
}
\arguments{
\item{\code{x}}{A data frame of numeric values.}
}
\value{
\code{multiprod} returns a data frame of numeric values with X number of columns, where X = number of phenotpyic traits.
}
\description{
Generates the product of all pairwise combinations of variables. xi * xi products are labeled with VariableName.Sq, whereas xi * xj are labeled with VariableNamei x VariableNamej.
}
\examples{
data(BumpusMales)

multiprod(BumpusMales[,3:11])
}

