% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseadd.R
\name{pairwisereplace}
\alias{pairwisereplace}
\title{Iteratively replace components with scaled components, and rescale to a composition.}
\usage{
pairwisereplace(weight1, weight2, component = rep(1, ncol(weight1)))
}
\arguments{
\item{weight1, weight2}{A matrix specifying how the mass in the component is split.}

\item{component}{A vector of integers specifying the component to be split each time.}
}
\description{
Replaces the mass in a component with two new components that are multiples of the original component, with the \code{weight1} the multiple for the first new component and \code{weight2} the multiple for the second new component
The compositional data has the same number of rows as \code{weight} and begins as all \code{1}, with one component.
}
