% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisesplit.R
\name{pairwisesubpartition}
\alias{pairwisesubpartition}
\title{Iteratively split components by given weight.}
\usage{
pairwisesubpartition(weight, component = rep(1, ncol(weight)))
}
\arguments{
\item{weight}{A matrix specifying how the mass in the component is split.}

\item{component}{A vector of integers specifying the component to be split each time.}
}
\description{
Splits the mass in a component into two, with the \code{weight} dictating how much of the mass goes to the first new component.
The compositional data has the same number of rows as \code{weight} and begins as all \code{1}, with one component.
}
