% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchBalance-plot.R
\name{plot.MatchBalance}
\alias{plot.MatchBalance}
\title{Plot method for MatchBalance.}
\usage{
\method{plot}{MatchBalance}(x, min.label.gap = 0.05, importance = TRUE,
  importance.sep = ifelse(ncol(x$df.matrix) > 10, " ", "\\n"), ...)
}
\arguments{
\item{x}{results of \code{\link{MatchBalance}}.}

\item{min.label.gap}{the minimum gap between x-axis labels on the top panel.}

\item{importance}{include importance metric (t-statistic) in axis labels.}

\item{importance.sep}{a character used to separate the covariate label and importance metric.}

\item{...}{currently unused.}
}
\value{
Returns a list invisibly with ggplot2 objects of the three panels.
}
\description{
Plot method for MatchBalance.
}
