% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdendrogram2.R
\name{plotdendrogram2}
\alias{plotdendrogram2}
\title{Plot a dendrogram of the merges.}
\usage{
plotdendrogram2(
  result,
  plot = TRUE,
  edge.root = TRUE,
  edgePar = list(p.lty = 0, t.cex = 1, t.adj = c(1, 1)),
  horiz = TRUE,
  nodeheight = "step",
  colors = NULL,
  ...
)
}
\arguments{
\item{result}{The \code{dendrogram.input} value from \code{\link[=psa]{psa()}}.}

\item{plot}{logical; if TRUE, plot the dendrogram output.}

\item{edge.root}{logical; if TRUE, draw an edge to the root node.}

\item{edgePar}{a list of plotting parameters for edges. See \code{\link[graphics:segments]{graphics::segments()}}.}

\item{horiz}{logical; if TRUE, draw the dendrogram horizontally.}

\item{nodeheight}{a string specifying node height.
"step" for heights given by the number of vertices after the merge
or "rmse scores" for heights given by the RMSE of the scores of the merge.}

\item{colors}{a vector of colors of nodes}

\item{...}{Passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\description{
Converts results to a \code{dendrogram} (see \code{\link[stats:dendrogram]{stats::as.dendrogram()}}) object and plots it.
}
