% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchBalance-print.R
\name{print.MatchBalance}
\alias{print.MatchBalance}
\title{Print method for MatchBalance.}
\usage{
\method{print}{MatchBalance}(x, ...)
}
\arguments{
\item{x}{result of \code{\link{MatchBalance}}.}

\item{...}{currently unused.}
}
\description{
Print method for MatchBalance.
}
