% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectsuborthant.R
\name{projectsuborthant}
\alias{projectsuborthant}
\title{Project data to a lower suborthant}
\usage{
projectsuborthant(X, V = diag(ncol(X)), v1, v2, w)
}
\arguments{
\item{X}{a data matrix whose coordinates are computed with respect to \code{V}.}

\item{V}{matrix for vertices of the reduced simplex.}

\item{v1, v2}{vertices to be merged.}

\item{w}{weight for merge.}
}
\value{
\itemize{
\item \code{X} projected data points in the new coordinates
\item \code{V} new vertices
\item \code{scores} Signed scores of the projection
\item \code{mergedirection} The direction of the merge, with positive pointing towards positive score points.
}
}
\description{
Project data to a lower suborthant
}
\details{
The order of v1 and v2 specify the sign of the scores: points on the v2 side of the orthant have positive scores
}
