% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-package.R
\docType{data}
\name{psa_citations}
\alias{psa_citations}
\title{Number of articles related to propensity score analysis in the Web of Science
and Google Scholar database.}
\format{a data frame with 92 observations of 4 variables.}
\description{
Number of articles related to propensity score analysis in the Web of Science
and Google Scholar database.
}
\keyword{datasets}
