% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_shiny.R
\name{psa_shiny}
\alias{psa_shiny}
\title{Run the PSA shiny application.}
\usage{
psa_shiny(...)
}
\arguments{
\item{...}{other parameters passed to [shiny::runApp].}
}
\description{
Run the PSA shiny application.
}
