% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqunif_plot.R
\name{qqunif.plot}
\alias{qqunif.plot}
\title{PP plot GWAS}
\usage{
qqunif.plot(pvalues, pvalues2 = NULL, should.thin = T,
  thin.obs.places = 2, thin.exp.places = 2,
  xlab = expression(paste("Expected (", -log[10], " p-value)")),
  ylab = expression(paste("Observed (", -log[10], " p-value)")),
  draw.conf = TRUE, conf.points = 1000, conf.col = "lightgray",
  conf.alpha = 0.05, pch = 20, aspect = "iso",
  prepanel = prepanel.qqunif, par.settings = list(superpose.symbol =
  list(pch = pch)), ...)
}
\arguments{
\item{pvalues}{list of p-values output}

\item{pvalues2}{specified when plot observed vs observed}
}
\description{
https://genome.sph.umich.edu/wiki/Code_Sample:_Generating_QQ_Plots_in_R
}
