% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoData.R
\name{readGitHub}
\alias{readGitHub}
\title{Read raw data files that are uploaded on GitHub}
\usage{
readGitHub(url, header = TRUE, fill = TRUE, stringAsFactors = FALSE, ...)
}
\arguments{
\item{\code{url}}{Character string containing the HTML url for the raw data. Supports the same file formats as \code{read.table}.}

\item{\code{header}}{Logical. Indicate whether header lines should be included. Default is \code{TRUE}.}

\item{\code{fill}}{Logical. Indicate whether blank fields in rows should be filled with "NA" values to insure that rows have equal length. Default is \code{TRUE}.}

\item{\code{stringsAsFactors}.}{Logical. Indicate whether character vectors should be automatically converted to factors. Default is \code{FALSE}.}

\item{\code{...}}{Further arguments that can be passed to the read.table family}
}
\description{
\code{readGitHub} allows one to read raw data from GitHub repositories
}
\details{
See documentation under ?read.table for specific information regarding data import.
}
\examples{
# Load the Bumpus.csv data file from GitHub
FlowersData <- readGitHub("https://raw.githubusercontent.com/MorphoFun/tremors/master/dataraw/Baranzelli_etal_Flowers.csv", sep = ",")

# Review the first 6 rows
head(FlowersData)
}
\seealso{
\code{read.table}, \code{connections}
}

