% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurse_tolowerdimension.R
\name{recurse_tolowerdimension}
\alias{recurse_tolowerdimension}
\title{Recurse through dimensions}
\usage{
recurse_tolowerdimension(
  X,
  V = diag(ncol(X)),
  evaluator,
  projector,
  testweights = seq(0, 1, length.out = 100),
  merges = NULL,
  maxsteps = NULL
)
}
\arguments{
\item{X}{matrix of data points for a simplex with vertices at the basis coordinates. Each row an observation.}

\item{V}{matrix of vertices in the basis of the \emph{original} data. Each row a vertex corresponding to the \emph{columns} of \code{X}.}

\item{evaluator}{A function that evaluates each proposed merge and return a single non-negative value for the quality of the merge. Arguments of \code{X}, \code{v1}, \code{v2}, and \code{w} at least}

\item{projector}{A function that takes result of \code{gridsearch()} and applies a projection. Returning \code{X}, \code{V}, \code{scores}, \code{mergedirection}.}

\item{testweights}{A vector of weights to try.}

\item{merges}{if non-NULL, then a data frame of prespecified merges.}

\item{maxsteps}{maximum number of merges to compute.}
}
\description{
Given a merge \code{evaluator} and a merge \code{projector}, recurse through dimensions keeping track of the appropriate data.
}
\details{
\itemize{
\item \code{Vbirth} keeps track of the merge that created each vertex. The merge is recorded acording to the destination dimension of the merge, so a value of \code{i} means the merge that created that vertex was creating the data for dimension \code{d - (i-1)}, where \code{d} is the starting dimension.
}
}
