% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_pca.R
\name{replace_zero}
\alias{replace_zero}
\title{Replace zeros by a fraction of columnwise minimum of nonzero observations}
\usage{
replace_zero(X, alpha = 0.5)
}
\arguments{
\item{X}{a data matrix. Columns are features and rows are samples.}

\item{alpha}{a positive scalar between 0 and 1. Zeros are replaced by alpha times
the minimum of nonzero observations}
}
\description{
Replace zeros by a fraction of columnwise minimum of nonzero observations
}
