% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification_balance.R
\name{stratification_balance}
\alias{stratification_balance}
\alias{summary.stratification_balance}
\alias{plot.stratification_balance}
\title{Calculate the balance across covariates for varying stratification sizes.}
\usage{
stratification_balance(
  covariates,
  treatment,
  ps,
  standardize = TRUE,
  n_strata = 1:10
)

\method{summary}{stratification_balance}(object, ...)

\method{plot}{stratification_balance}(x, ...)
}
\arguments{
\item{covariates}{either a vector (for one covariate) or data.frame.}

\item{treatment}{vector for the treatment (should be logical or integer).}

\item{ps}{propensity scores.}

\item{standardize}{if TRUE, the covariates will be standardized before calculating differences
(i.e. converted to z-scores).}

\item{n_strata}{how many strata to estimate the balance.}

\item{object}{result from [stratification_balance()].}

\item{...}{currently not used.}

\item{x}{result from [stratification_balance()].}
}
\value{
a data frame with the average difference/balance for each covariate and stratification size.

a ggplot2 expression.
}
\description{
Cochran (1968) suggested that by stratifying with 2, 3, 4, 5, and 6 strata would reduce bias by 
64%, 79%, 86%, 90%, and 92% respectively. This function, along with the accompanying `summary`
and `plot` functions calculates the mean difference between treatment and control groups across
all strata for varying number of stratifications.
}
\examples{
# Create a simulated dataset
set.seed(2112) 
n <- 1000
treatment_effect <- 1.5
X <- mvtnorm::rmvnorm(n,
 					 mean = c(0.5, 1, 0),
 					 sigma = matrix(c(2, 1, 1,
 					   				  1, 1, 1,
 					   				  1, 1, 1), ncol = 3) )
dat <- tibble::tibble(
	 x1 = X[, 1],
	 x2 = X[, 2],
	 x3 = X[, 3] > 0,
	 treatment = as.numeric(- 0.5 + 0.25 * x1 + 0.75 * x2 + 0.05 * x3 + rnorm(n, 0, 1) > 0),
	 outcome = treatment_effect * treatment + rnorm(n, 0, 1)
)

# Estimate the propensity scores
lr_out <- glm(treatment ~ x1 + x2 + x3, data = dat, family = binomial(link = 'logit'))

diff_out <- stratification_balance(
	covariates = dat[,1:3],
	treatment = dat$treatment,
	ps = fitted(lr_out)
)

# Note that any place where the mean_difference is NA indicates that one of the groups
# did not have any observations. This will occur when the number of stata increase.
summary(diff_out) 

cochran_bias_reduction <-  c(.64, .79, .86, .90)
plot(diff_out) +
 	geom_hline(yintercept = 1 - cochran_bias_reduction, linetype = 2) +
 	annotate(geom = 'text', label = paste0(100 * cochran_bias_reduction, '\%'), 
 			 x = 1, y = 1 - cochran_bias_reduction,
 			 vjust = -0.2, hjust = -0.1, size = 3)

}
\references{
Cochran, W.G. (1968). The effectiveness of adjustment by subclassification in 
removing bias in observational studies. *Biometrics. 24*(2). pp 295-313. 
<https://doi.org/10.2307/2528036>
}
