% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_modes.R
\name{tidy_modes}
\alias{tidy_modes}
\title{Prepare modes of variation for ggplotting}
\usage{
tidy_modes(result, ..., dimensions = NULL)
}
\arguments{
\item{result}{a list generated by \code{\link[=psa]{psa()}}.}

\item{...}{Data frames of covariates with rows corresponding exactly to the input data. Column names should not be the same as column names of the input.}

\item{dimensions}{Supply the number of vertices after the merge corresponding to the modes of variation. By default all dimensions are included.}
}
\description{
Prepare modes of variation for ggplotting
}
