% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolowerdimension.R
\name{tolowerdimension}
\alias{tolowerdimension}
\title{Finds and applies best merge}
\usage{
tolowerdimension(
  X,
  V = diag(1, ncol(X)),
  evaluator,
  projector,
  testweights = seq(0, 1, length.out = 100),
  merge = NULL
)
}
\arguments{
\item{X}{matrix of data points for a simplex with vertices at the basis coordinates. Each row an observation.}

\item{V}{matrix of vertices in the basis of the \emph{original} data. Each row a vertex corresponding to the \emph{columns} of \code{X}.}

\item{evaluator}{A function that evaluates each proposed merge and return a single non-negative value for the quality of the merge. Arguments of \code{X}, \code{v1}, \code{v2}, and \code{w} at least}

\item{projector}{A function that takes result of \code{gridsearch()} and applies a projection. Returning \code{X}, \code{V}, \code{scores}, \code{mergedirection}.}

\item{testweights}{A vector of weights to try.}

\item{merge}{If supplied, used instead of \code{\link[=gridsearch]{gridsearch()}}. \code{bestmerge} must have same structure as result from \code{\link[=gridsearch]{gridsearch()}}}
}
\value{
A list
\itemize{
\item \code{X} Projected data points in with the new vertices as the coordinates
\item \code{V} New set of vertices with row names.
\item \code{scores} Signed scores for the merge
\item \code{mergedirection} The unit vector parallel to \code{v2 - v1} and in the direction of positive scores, and in the original coordinates.
\item \code{v1} The index of \code{v1} used in \code{projector} (corresponding to a row of the old vertices matrix).
\item \code{v2} The index of \code{v2} used in \code{projector} (corresponding to a row of the old vertices matrix).
\item \code{w} The weight of the merge between \code{v1} and \code{v2}. As in \code{w * v1 + (1-w) * v2}, where \code{v1} and \code{v2} are the first and second vertices given by \code{mergedvertices}.
}
}
\description{
Finds and applies best merge
}
\details{
If \code{merge} is not supplied and the \emph{largest} score of the found merge is \emph{negative}, then swaps the signs of the scores and records this by swapping the returned \code{v1} with \code{v2}, changing w to 1-w, and flipping the direction of \code{mergedirection}. This should mean that the returned merge information can be reapplied on new data and the direction will be the same.

If \code{merge} is non-\code{NULL} then \code{evaluator} is ignored - no search occurs.
}
