% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-package.R
\docType{data}
\name{tutoring}
\alias{tutoring}
\title{Data from a study examining the effects of tutoring on English grades.}
\format{
a data frame with 1,381 observations of 17 variables.
}
\description{
Data from a study examining the effects of tutoring on English grades.
}
\keyword{datasets}
