% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2rds.R
\name{txt2rds}
\alias{txt2rds}
\title{txt file to RDS file}
\usage{
txt2rds(i, input, output)
}
\arguments{
\item{i}{chromosome indicator, one of 1, ..., 23, "X"; for print purpose only}

\item{input}{input txt file}

\item{output}{output Rds file}
}
\description{
txt file to RDS file
}
