% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_plot.R
\name{weighting_plot}
\alias{weighting_plot}
\title{Propensity score weighting plot}
\usage{
weighting_plot(
  ps,
  treatment,
  outcome,
  estimand = "ATE",
  null_hypothesis = 0,
  method = "loess",
  se = TRUE,
  level = 0.95,
  colors = c("#fc8d62", "#66c2a5"),
  point_size_range = c(1, 8),
  xlab = "Propensity Score",
  ylab = "Outcome",
  treat_lab = "Treatment",
  size_lab = "PS Weight",
  plot_points = TRUE
)
}
\arguments{
\item{ps}{vector of propensity scores.}

\item{treatment}{vector of treatment indicators.}

\item{outcome}{vector of outcome values.}

\item{estimand}{the estimand to use, either ATE, ATT, ATC, or ATM.}

\item{null_hypothesis}{the value of the null hypothesis (typically zero). A
horizontal line will be drawn to compare with regression line.}

\item{method}{the method to use for the regression line. Typically loess, gan, or lm.}

\item{se}{Display confidence interval around smooth? (TRUE by default, see level to control.)}

\item{level}{Level of confidence interval to use (0.95 by default).}

\item{colors}{vector of colors to use for control and treatment, respectively.}

\item{point_size_range}{range of the point sizes. Needs to be a vector of length two.}

\item{xlab}{label for the x-axis.}

\item{ylab}{label for the y-axis.}

\item{treat_lab}{label for the legend of group colors.}

\item{size_lab}{label for the legend of point sizes.}

\item{plot_points}{whether to plot the individual points.}
}
\value{
a ggplot2 expression.
}
\description{
Propensity score weighting plot
}
\examples{
if(require(Matching)) {
data(lalonde, package = 'Matching')
lr_out <- glm(treat ~ age + I(age^2) + educ + I(educ^2) + black + 
              hisp + married + nodegr + re74  + I(re74^2) + re75 + I(re75^2) +
              u74 + u75,
			  data = lalonde, 
			  family = binomial(link = 'logit'))
lalonde$ps <- fitted(lr_out)
weighting_plot(ps = lalonde$ps,
					treatment = lalonde$treat,
					outcome = log(lalonde$re78))
}
}
