% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-package.R
\docType{data}
\name{wos}
\alias{wos}
\title{Number of articles related to propensity score analysis in the Web of Science
database.}
\format{a data frame with 20 observations of 2 variables.}
\description{
Number of articles related to propensity score analysis in the Web of Science
database.
}
\keyword{datasets}

