% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionDifferentials.R
\name{dCompare}
\alias{dCompare}
\title{Comparing estimates of selection differentials}
\usage{
dCompare(w, z)
}
\arguments{
\item{\code{w}}{Relative fitness.}

\item{\code{z}}{Phenotypic trait(s). Character values are not accepted.}
}
\value{
\code{dCompare} returns a matrix of numeric values with 7 rows and X columns, where X = number of phenotpyic traits.
}
\description{
\code{dCompare} allows the user to evaluate how different methods for calculating phenotypic section differentials influence the output and, thus, the interpretation of indirect selection on phenotypic traits.
}
\section{Output}{
  \code{dCov} is the selection differential calculated from the covariance between the relative fitness, \code{w}, and each phenotypic trait \code{z} (Lande and Arnold 1983).

  \code{dMean} is the selection differential calculated as  as described in Brodie et al. (1995). Not output for Gaussian fitness measures.

  \code{dMean_stdsd} is the selection differential calculated as described in Gvoždík and Smolinský (2015). Not output for Gaussian fitness measures.

  \code{dReg} is the selection differential calculated as the partial regression coefficient of relative fitness, \code{w}, against each individual phenotypic trait through univariate linear regressions.

  \code{dCovScale} calculates the selection differential using the equation in \code{dCovScale}, but uses z data that are standardized to a mean of zero and unit variance.

  \code{dMeanScale} calculates the selection differential using the equation in \code{dMeanScale}, but uses z and z* data that are standardized to a mean of zero and unit variance. Not output for Gaussian fitness measures.

  \code{dRegScale} calculates the selection differential using the equation in \code{dRegScale}, but uses z data that are standardized to a mean of zero and unit variance.
}
\examples{
data(BumpusMales)

dCompare(BumpusMales$w, BumpusMales[,3:11])
}
\references{
Brodie III ED, Moore AJ, Janzen FJ. 1995. Visualizing and quantifying natural selection. \emph{Trends in Ecology and Evolution} 10(8): 313-318. \url{http://www.sciencedirect.com/science/article/pii/S016953470089117X}

Gvoždík L, Smolinský R. 2015. Body size, swimming speed, or thermal sensitivity? Predator-imposed selection on amphibian larvae. \emph{BMC Evolutionary Biology} 15: 238. \url{http://bmcevolbiol.biomedcentral.com/articles/10.1186/s12862-015-0522-y}

Lande R, Arnold SJ. 1983. The measurement of selection on correlated characters. \emph{Evolution} 37(6): 1210-1226. \url{http://www.jstor.org/stable/2408842}
}

