% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.mids.R
\name{merge.mids}
\alias{merge.mids}
\title{Merge imputed dataset with original data.frame.}
\usage{
\method{merge}{mids}(x, y, shadow.matrix = FALSE, shadow.suffix = "_missing", ...)
}
\arguments{
\item{x}{the results from \code{\link{mice::mice}}.}

\item{y}{the \code{data.frame} passed to \code{\link{mice::mice}}. This must
have the same number rows, but may contain additional columns.}

\item{shadow.matrix}{if TRUE, addtional columns will be added indicating 
whether the value was missing in the original data.frame.}

\item{shadow.suffix}{the suffix used for the shadow matrix.}

\item{...}{parameters passed to \code{\link{mice::complete}}.}
}
\description{
Returns a data.frame where the variables in the original (i.e. \code{y} 
parameter) are replaced by the imputed variables from mice (i.e. \code{x}
parameter).
}
