% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-package.R
\docType{package}
\name{psa-package}
\alias{psa-package}
\title{Applied Propensity Score Analysis with R}
\description{
Applied Propensity Score Analysis with R
}
\author{
\email{jason@bryer.org}
}
\keyword{analysis}
\keyword{matching}
\keyword{package}
\keyword{propensity}
\keyword{psa}
\keyword{score}
