% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchBalance-summary.R
\name{summary.MatchBalance}
\alias{summary.MatchBalance}
\title{Summary method for MatchBalance.}
\usage{
\method{summary}{MatchBalance}(object, print = TRUE, ...)
}
\arguments{
\item{object}{result from \code{\link{MatchBalance}}}

\item{print}{whether to print the result or to return using \code{invisible}.}

\item{...}{currently unused.}
}
\description{
Summary method for MatchBalance.
}
