% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{treatment_effect}
\alias{treatment_effect}
\title{Estimate the treatment effects.}
\usage{
treatment_effect(treatment, outcome, ps, weights, ...)
}
\arguments{
\item{treatment}{logical vector for treatment status.}

\item{outcome}{vector of outcome values.}

\item{ps}{vector of propensity scores.}

\item{weights}{vector of propensity score weights.}

\item{...}{parameters passed [calculate_ps_weights()].}
}
\value{
the treatment effect.
}
\description{
Estimate the treatment effects.
}
